Modify PDF Layout
Added logo and watermark

Modify PDF Layout

I want to use image in pdf header on top left instead of text

I dont want to use svg replave it with image

I want to use a image watermark for the pdf. Ensure that it is not distu

okay apply this method and ensure watermark is behind the texts and the

what is this error

the log in the header is too small, Make the header logo big. the waterm

new error

when I try to download pdf

Remove the diagonal watermark make it simple as before

but the header logo on top left is still very small

still very small. Also how to move it more towards the left so it utiliz

Modify PDF Layout